move jwt from common to core functions
