+++ common/lib/idds/common/authentication.py
@@ -11,7 +11,6 @@
 import datetime
 import base64
 import json
-import jwt
 import os
 import re
 import requests
@@ -31,12 +30,7 @@ except ImportError:
     from urllib.parse import urlencode
     raw_input = input
 
-# from cryptography import x509
-from cryptography.hazmat.primitives.asymmetric.rsa import RSAPublicNumbers
-from cryptography.hazmat.backends import default_backend
-from cryptography.hazmat.primitives import serialization
-
-# from idds.common import exceptions
+from idds.common import exceptions
 from idds.common.constants import HTTP_STATUS_CODE
 
 
@@ -285,58 +279,10 @@ class OIDCAuthentication(BaseAuthentication):
             return False, 'Failed to refresh oidc token: ' + str(error)
 
     def get_public_key(self, token, jwks_uri, no_verify=False):
-        headers = jwt.get_unverified_header(token)
-        if headers is None or 'kid' not in headers:
-            raise jwt.exceptions.InvalidTokenError('cannot extract kid from headers')
-        kid = headers['kid']
-
-        jwks = self.get_cache_value(jwks_uri)
-        if not jwks:
-            jwks_content = self.get_http_content(jwks_uri, no_verify=no_verify)
-            jwks = json.loads(jwks_content)
-            self.set_cache_value(jwks_uri, jwks)
-
-        jwk = None
-        for j in jwks.get('keys', []):
-            if j.get('kid') == kid:
-                jwk = j
-        if jwk is None:
-            raise jwt.exceptions.InvalidTokenError('JWK not found for kid={0}: {1}'.format(kid, str(jwks)))
-
-        public_num = RSAPublicNumbers(n=decode_value(jwk['n']), e=decode_value(jwk['e']))
-        public_key = public_num.public_key(default_backend())
-        pem = public_key.public_bytes(encoding=serialization.Encoding.PEM, format=serialization.PublicFormat.SubjectPublicKeyInfo)
-        return pem
+        raise exceptions.NotImplementedException("Method get_public_key is not implemented.")
 
     def verify_id_token(self, vo, token):
-        try:
-            auth_config, endpoint_config = self.get_auth_endpoint_config(vo)
-
-            # check audience
-            decoded_token = jwt.decode(token, verify=False, options={"verify_signature": False})
-            audience = decoded_token['aud']
-            if audience not in [auth_config['audience'], auth_config['client_id']]:
-                # discovery_endpoint = auth_config['oidc_config_url']
-                return False, "The audience %s of the token doesn't match vo configuration(client_id: %s)." % (audience, auth_config['client_id']), None
-
-            public_key = self.get_public_key(token, endpoint_config['jwks_uri'], no_verify=auth_config['no_verify'])
-            # decode token only with RS256
-            if 'iss' in decoded_token and decoded_token['iss'] and decoded_token['iss'] != endpoint_config['issuer'] and endpoint_config['issuer'].startswith(decoded_token['iss']):
-                # iss is missing the last '/' in access tokens
-                issuer = decoded_token['iss']
-            else:
-                issuer = endpoint_config['issuer']
-
-            decoded = jwt.decode(token, public_key, verify=True, algorithms='RS256',
-                                 audience=audience, issuer=issuer)
-            decoded['vo'] = vo
-            if 'name' in decoded:
-                username = decoded['name']
-            else:
-                username = None
-            return True, decoded, username
-        except Exception as error:
-            return False, 'Failed to verify oidc token: ' + str(error), None
+        raise exceptions.NotImplementedException("Method verify_id_token is not implemented.")
 
     def setup_oidc_client_token(self, issuer, client_id, client_secret, scope, audience):
         try:
